Child is extracted
