+++ kotlin-react/src/main/kotlin/react/React.kt
@@ -1,17 +1,6 @@
 package react
 
-import kotlinext.js.asJsObject
 import kotlinext.js.jsObject
-
-external interface Child
-
-fun Child.asStringOrNull(): String? =
-    asDynamic() as? String
-
-fun Child.asElementOrNull(): ReactElement? = when {
-    asJsObject().hasOwnProperty("\$\$typeof") -> unsafeCast<ReactElement>()
-    else -> null
-}
 
 external interface ReactElement : Child {
     val props: RProps
